⚰️ :: 수정 사항 반영 (Delete AppBar, Card, ViewModel)
2023-01-29
