--- app/src/main/java/kr/hs/dgsw/smartschool/dui/viewmodel/MainViewModel.kt
@@ -1,5 +1,0 @@
-package kr.hs.dgsw.smartschool.dui.viewmodel
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel : ViewModel()
